+++ app/controllers/admin/contents_controller.rb
@@ -1,5 +1,5 @@
 module Admin
-  class ContentsController < ::Admin::BaseController
+  class ContentsController < Spree::Admin::BaseController
     def edit
       @preference_sections = preference_sections.map do |preference_section|
         { name: preference_section.name, preferences: preference_section.preferences }
