+++ app/controllers/admin/stripe_connect_settings_controller.rb
@@ -1,7 +1,7 @@
 # This controller is used by super admin users to update the settings the app is using
 
 module Admin
-  class StripeConnectSettingsController < ::Admin::BaseController
+  class StripeConnectSettingsController < Spree::Admin::BaseController
     StripeConnectSettings = Struct.new(:stripe_connect_enabled)
 
     before_action :load_settings, only: [:edit]
