+++ app/controllers/spree/admin/orders/customer_details_controller.rb
@@ -1,7 +1,7 @@
 module Spree
   module Admin
     module Orders
-      class CustomerDetailsController < ::Admin::BaseController
+      class CustomerDetailsController < Spree::Admin::BaseController
         before_action :load_order
         before_action :check_authorization
         before_action :set_guest_checkout_status, only: :update
