+++ app/controllers/spree/admin/payments_controller.rb
@@ -2,7 +2,7 @@
 
 module Spree
   module Admin
-    class PaymentsController < ::Admin::BaseController
+    class PaymentsController < Spree::Admin::BaseController
       before_action :load_order, except: [:show]
       before_action :load_payment, only: [:fire, :show]
       before_action :load_data
