+++ app/controllers/spree/admin/search_controller.rb
@@ -1,6 +1,6 @@
 module Spree
   module Admin
-    class SearchController < ::Admin::BaseController
+    class SearchController < Spree::Admin::BaseController
       # http://spreecommerce.com/blog/2010/11/02/json-hijacking-vulnerability/
       before_action :check_json_authenticity, only: :index
       respond_to :json
