+++ app/controllers/spree/admin/tax_settings_controller.rb
@@ -1,6 +1,6 @@
 module Spree
   module Admin
-    class TaxSettingsController < ::Admin::BaseController
+    class TaxSettingsController < Spree::Admin::BaseController
       def update
         Spree::Config.set(params[:preferences])
