+++ app/controllers/spree/admin/taxons_controller.rb
@@ -1,6 +1,6 @@
 module Spree
   module Admin
-    class TaxonsController < ::Admin::BaseController
+    class TaxonsController < Spree::Admin::BaseController
       respond_to :html, :json, :js
 
       def create
