+++ engines/order_management/app/controllers/order_management/reports/enterprise_fee_summaries_controller.rb
@@ -2,7 +2,7 @@
 
 module OrderManagement
   module Reports
-    class EnterpriseFeeSummariesController < ::Admin::BaseController
+    class EnterpriseFeeSummariesController < Spree::Admin::BaseController
       before_filter :load_report_parameters
       before_filter :load_permissions
