+++ spec/features/admin/variant_overrides_spec.rb
@@ -196,7 +196,7 @@ feature "
             expect(page).to have_content "Changes to one override remain unsaved."
 
             # Set a user without suficient permissions
-            allow_any_instance_of(Admin::BaseController).to receive(:current_spree_user).and_return(build(:user))
+            allow_any_instance_of(Spree::Admin::BaseController).to receive(:current_spree_user).and_return(build(:user))
 
             expect do
               click_button 'Save Changes'
